fix whitespace handling to not reformat document
Program.cs (and the unit tests) called NormalizeWhitespace on the entire syntax tree, which has the affect of reformatting the entire document. It doesn't "pretty-print", it just removes all extra whitespace.

This leads to poorly formatted output. So now, we remove the global NormalizeWhitespace and replace it with smaller, targeted calls to new syntax nodes, to avoid reformatting the entire document.