convert TestCaseSource to MemberData
There was already some support for this in FieldVisitor, but it is incomplete, and does not handle when TestCaseSource or MemberData are methods. I replaced it with a ClassVisitor.cs

Additionally, the visitor was becoming huge, so I pulled out the separate conversions into different feature classes.

Finally, I found a bug where we remove XML doc comments when we remove the [TestFixture] class attribute, so I fixed that as per https://joshvarty.com/2014/08/15/learn-roslyn-now-part-5-csharpsyntaxrewriter/